requirements.txt
Flask
numpy
Pillow
transformers
paddleocr
torch
